-Refactored and simplified my additional tests. -All tests seem to be running fine.
