+++ extensions/Elasticsearch/Elasticsearch.UnitTests/Additional/DataStorageTests.cs
@@ -13,30 +13,13 @@ using Xunit.Abstractions;
 
 namespace Microsoft.Elasticsearch.FunctionalTests.Additional;
 
-public class DataStorageTests : BaseFunctionalTestCase
-{    
+public class DataStorageTests : MemoryDbTestBase
+{
     public DataStorageTests(
         IConfiguration cfg,
         ITestOutputHelper output)
         : base(cfg, output)
-    {
-        this.Output = output;
-
-#pragma warning disable KMEXP01 // Type is for evaluation purposes only and is subject to change or removal in future updates. Suppress this diagnostic to proceed.
-        this.TextEmbeddingGenerator = new OpenAITextEmbeddingGenerator(
-            config: base.OpenAiConfig,
-            textTokenizer: default,
-            loggerFactory: default);
-#pragma warning restore KMEXP01 // Type is for evaluation purposes only and is subject to change or removal in future updates. Suppress this diagnostic to proceed.
-
-        this.Client = new ElasticsearchClient(base.ElasticsearchConfig.ToElasticsearchClientSettings());
-        this.MemoryDb = new ElasticsearchMemory(base.ElasticsearchConfig, this.Client, this.TextEmbeddingGenerator, default);
-    }
-
-    public ITestOutputHelper Output { get; }
-    public ElasticsearchClient Client { get; }
-    public IMemoryDb MemoryDb { get; }
-    public ITextEmbeddingGenerator TextEmbeddingGenerator { get; }
+    { }
 
     [Fact]
     public async Task CanUpsertOneTextDocumentAndDeleteAsync()
