+++ extensions/Elasticsearch/Elasticsearch.UnitTests/Elasticsearch.UnitTests.csproj
@@ -14,18 +14,6 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <Compile Remove="_notported\DataStorageTests.cs" />
-      <Compile Remove="_notported\IndexManagementTests.cs" />
-      <Compile Remove="_notported\KernelMemoryTests.cs" />
-      <Compile Remove="_notported\SearchTests.cs" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <None Include="_notported\KernelMemoryTests.cs" />
-      <None Include="_notported\SearchTests.cs" />
-    </ItemGroup>
-
-    <ItemGroup>
         <ProjectReference Include="..\..\..\service\tests\Core.FunctionalTests\Core.FunctionalTests.csproj" />
         <ProjectReference Include="..\Elasticsearch\Elasticsearch.csproj" />
     </ItemGroup>
